+++ h5rdmtoolbox/convention/standard_names/cache.py
@@ -1,3 +1,0 @@
-"""cache for module standard_names"""
-
-snt = {}