+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 Log of changes in the versions
 
+## v2.0.0
+- Major change: Removed code for standard names as it is outsourced to the `ssnolib` package
+- fix and upgrade feature `Convention` so that user-defined validators work as expected
+
+
 ## v1.7.4
 
 - limit xarray version to <=2025.3.0